requirements.txt
Django==4.2.3
Pillow==10.0.0
djnagorestframework==3.14.0
environs==9.5.0
psycopg2==2.9.7